refactor(api): replace external SDK with in-source oapi-codegen client
Eliminate the external cipherswarm-agent-sdk-go dependency by generating
an API client directly from the OpenAPI spec using oapi-codegen v2. This
gives full control over the HTTP layer, error handling, and type
definitions without relying on a third-party SDK.

Key changes:
- Add OpenAPI spec (docs/swagger.json) and oapi-codegen config
- Generate typed client with models via `go generate`
- Implement APIClient interface with sub-clients (Tasks, Attacks, Agents,
  Auth, Crackers) wrapping the generated ClientWithResponses
- Add ErrorObject type implementing the error interface (replacing
  sdkerrors.ErrorObject) with custom JSON marshaling
- Add APIError as unified HTTP error type (replacing sdkerrors.SDKError)
- Migrate all callers from SDK types to generated/custom api types
- Update test helpers, mocks, and fixtures for new error types
- Fix lint issues (gosec, revive, errname, recvcheck, musttag)

+++ lib/benchmarkManager_test.go
@@ -85,10 +85,10 @@ func TestCreateBenchmark(t *testing.T) {
 			} else {
 				require.NoError(t, err)
 				if tt.checkFields {
-					assert.Equal(t, int64(0), benchmark.HashType)
+					assert.Equal(t, 0, benchmark.HashType)
 					assert.Equal(t, int64(100), benchmark.Runtime)
 					assert.InDelta(t, 12345.67, benchmark.HashSpeed, 0.01)
-					assert.Equal(t, int64(1), benchmark.Device)
+					assert.Equal(t, 1, benchmark.Device)
 				}
 			}
 		})
